refactor(dev-infra): move GitClient to common util
Moves GitClient from merge script into common utils for unified
method of performing git actions throughout the ng-dev toolset.
