+++ dev-infra/pr/merge/config.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {getConfig, NgDevConfig} from '../../utils/config';
+import {getConfig, GitClientConfig, NgDevConfig} from '../../utils/config';
 
 import {GithubApiMergeStrategyConfig} from './strategies/api-merge';
 
@@ -31,22 +31,12 @@ export interface TargetLabel {
   branches: string[]|((githubTargetBranch: string) => string[]);
 }
 
-/** Describes the remote used for merging pull requests. */
-export interface MergeRemote {
-  /** Owner name of the repository. */
-  owner: string;
-  /** Name of the repository. */
-  name: string;
-  /** Whether SSH should be used for merging pull requests. */
-  useSsh?: boolean;
-}
-
 /**
  * Configuration for the merge script with all remote options specified. The
  * default `MergeConfig` has does not require any of these options as defaults
  * are provided by the common dev-infra github configuration.
  */
-export type MergeConfigWithRemote = MergeConfig&{remote: MergeRemote};
+export type MergeConfigWithRemote = MergeConfig&{remote: GitClientConfig};
 
 /** Configuration for the merge script. */
 export interface MergeConfig {
@@ -54,7 +44,7 @@ export interface MergeConfig {
    * Configuration for the upstream remote. All of these options are optional as
    * defaults are provided by the common dev-infra github configuration.
    */
-  remote?: Partial<MergeRemote>;
+  remote?: GitClientConfig;
   /** List of target labels. */
   labels: TargetLabel[];
   /** Required base commits for given branches. */
