+++ dev-infra/utils/BUILD.bazel
@@ -7,6 +7,7 @@ ts_library(
     visibility = ["//dev-infra:__subpackages__"],
     deps = [
         "@npm//@octokit/graphql",
+        "@npm//@octokit/rest",
         "@npm//@types/inquirer",
         "@npm//@types/node",
         "@npm//@types/shelljs",
